api/fields.py
'''
Module with custom field types for serializers.
'''

import base64

from django.core.files.base import ContentFile
from rest_framework import serializers


class Base64ImageField(serializers.ImageField):
    '''
    Customizes 'to_internal_value' function
    to decode base64 image string into file.
    Returns image file address for GET requests.
    '''

    def to_internal_value(self, data):
        if isinstance(data, str) and data.startswith('data:image'):
            format, imgstr = data.split(';base64,')
            ext = format.split('/')[-1]
            data = ContentFile(base64.b64decode(imgstr), name='temp.' + ext)

        return super().to_internal_value(data)
