requirements.txt
Django==2.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-yasg==1.21.5
Pillow==8.3.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyJWT==2.1.0
pyyaml==6.0
requests==2.26.0
